+++ src/components/index.js
@@ -1,4 +1,4 @@
-export { default as GlobalMenu } from "./GlobalMenu";
+export { default as GlobalMenu } from "../scenes/global/GlobalMenu";
 export { default as HeadAndBody } from "./HeadAndBody";
 export { default as Header, PaddedContainer } from "./Header";
 export { default as VersionWarning } from "./VersionWarning";
